extracted GraphQLUserGroup from GraphQLUser
