+++ mockingbird/src/main/kotlin/com/leftindust/mockingbird/dao/PermissionDao.kt
@@ -2,7 +2,7 @@ package com.leftindust.mockingbird.dao
 
 import com.leftindust.mockingbird.auth.MediqToken
 import com.leftindust.mockingbird.dao.entity.AccessControlList
-import com.leftindust.mockingbird.graphql.types.GraphQLUser
+import com.leftindust.mockingbird.graphql.types.GraphQLUserGroup
 import com.leftindust.mockingbird.graphql.types.input.GraphQLPermissionInput
 
 interface PermissionDao {
@@ -13,7 +13,7 @@ interface PermissionDao {
     ): AccessControlList
 
     suspend fun addGroupPermission(
-        gid: GraphQLUser.Group.ID,
+        gid: GraphQLUserGroup.ID,
         permission: GraphQLPermissionInput,
         requester: MediqToken
     ): AccessControlList
