+++ mockingbird/src/main/kotlin/com/leftindust/mockingbird/graphql/mutations/GroupMutation.kt
@@ -1,20 +1,18 @@
 package com.leftindust.mockingbird.graphql.mutations
 
-import com.expediagroup.graphql.generator.scalars.ID
 import com.expediagroup.graphql.server.operations.Mutation
 import com.leftindust.mockingbird.auth.GraphQLAuthContext
 import com.leftindust.mockingbird.dao.GroupDao
-import com.leftindust.mockingbird.graphql.types.GraphQLPermission
-import com.leftindust.mockingbird.graphql.types.GraphQLUser
+import com.leftindust.mockingbird.graphql.types.GraphQLUserGroup
 import com.leftindust.mockingbird.graphql.types.input.GraphQLGroupInput
-import com.leftindust.mockingbird.graphql.types.input.GraphQLPermissionInput
 import org.springframework.stereotype.Component
 
 @Component
 class GroupMutation(private val groupDao: GroupDao) : Mutation {
-    suspend fun addGroup(group: GraphQLGroupInput, graphQLAuthContext: GraphQLAuthContext): GraphQLUser.Group {
+    // TODO: 2022-03-02 test this
+    suspend fun addGroup(group: GraphQLGroupInput, graphQLAuthContext: GraphQLAuthContext): GraphQLUserGroup {
         return groupDao
             .addGroup(group, graphQLAuthContext.mediqAuthToken)
-            .let { GraphQLUser.Group(it) }
+            .let { GraphQLUserGroup(it) }
     }
 }
