+++ mockingbird/src/main/kotlin/com/leftindust/mockingbird/graphql/mutations/PermissionMutation.kt
@@ -1,18 +1,17 @@
 package com.leftindust.mockingbird.graphql.mutations
 
-import com.expediagroup.graphql.generator.scalars.ID
 import com.expediagroup.graphql.server.operations.Mutation
 import com.leftindust.mockingbird.auth.GraphQLAuthContext
 import com.leftindust.mockingbird.dao.PermissionDao
 import com.leftindust.mockingbird.graphql.types.GraphQLPermission
-import com.leftindust.mockingbird.graphql.types.GraphQLUser
+import com.leftindust.mockingbird.graphql.types.GraphQLUserGroup
 import com.leftindust.mockingbird.graphql.types.input.GraphQLPermissionInput
 import org.springframework.stereotype.Component
 
 @Component
 class PermissionMutation(private val permissionDao: PermissionDao) : Mutation {
     suspend fun addPermission(
-        groupId: GraphQLUser.Group.ID? = null,
+        groupId: GraphQLUserGroup.ID? = null,
         userUid: String? = null,
         permission: GraphQLPermissionInput,
         graphQLAuthContext: GraphQLAuthContext
