+++ mockingbird/src/main/kotlin/com/leftindust/mockingbird/graphql/types/input/GraphQLUserInput.kt
@@ -4,13 +4,13 @@ import com.expediagroup.graphql.generator.annotations.GraphQLDescription
 import com.expediagroup.graphql.generator.annotations.GraphQLName
 import com.expediagroup.graphql.generator.execution.OptionalInput
 import com.leftindust.mockingbird.graphql.types.GraphQLDoctor
-import com.leftindust.mockingbird.graphql.types.GraphQLUser
+import com.leftindust.mockingbird.graphql.types.GraphQLUserGroup
 
 @GraphQLName("UserInput")
 data class GraphQLUserInput(
     val uid: String,
     val nameInfo: GraphQLNameInfoInput,
-    val group: GraphQLUser.Group.ID? = null,
+    val group: GraphQLUserGroup.ID? = null,
     val doctor: GraphQLDoctor.ID? = null,
 )
 
@@ -18,6 +18,6 @@ data class GraphQLUserInput(
 @GraphQLDescription("edits the user with the given uid, fields left unset will not be edited ")
 data class GraphQLUserEditInput(
     val uid: String,
-    val group: OptionalInput<GraphQLUser.Group.ID?> = OptionalInput.Undefined,
+    val group: OptionalInput<GraphQLUserGroup.ID?> = OptionalInput.Undefined,
     val doctor: OptionalInput<GraphQLDoctor.ID?> = OptionalInput.Undefined,
 )
